extract result-printing from hyperopt class
