+++ tests/commands/test_commands.py
@@ -920,7 +920,7 @@ def test_start_test_pairlist(mocker, caplog, tickers, default_conf, capsys):
 
 def test_hyperopt_list(mocker, capsys, caplog, hyperopt_results):
     mocker.patch(
-        'freqtrade.optimize.hyperopt.Hyperopt.load_previous_results',
+        'freqtrade.optimize.hyperopt_tools.HyperoptTools.load_previous_results',
         MagicMock(return_value=hyperopt_results)
     )
 
@@ -1152,7 +1152,7 @@ def test_hyperopt_list(mocker, capsys, caplog, hyperopt_results):
 
 def test_hyperopt_show(mocker, capsys, hyperopt_results):
     mocker.patch(
-        'freqtrade.optimize.hyperopt.Hyperopt.load_previous_results',
+        'freqtrade.optimize.hyperopt_tools.HyperoptTools.load_previous_results',
         MagicMock(return_value=hyperopt_results)
     )
